Refactor folder structures
