+++ lib/screens/settings/about/legal_modal.dart
