+++ lib/screens/settings/app_settings/advanced_settings/app_unlock_setup_modal.dart
@@ -5,8 +5,8 @@ import 'package:local_auth/local_auth.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
-import 'package:pi_hole_client/screens/settings/advanced_settings/app_lock/remove_passcode_modal.dart';
-import 'package:pi_hole_client/screens/settings/advanced_settings/app_lock/create_pass_code_modal.dart';
+import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/app_lock/remove_passcode_modal.dart';
+import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/app_lock/create_pass_code_modal.dart';
 
 import 'package:pi_hole_client/providers/app_config_provider.dart';
 import 'package:pi_hole_client/functions/snackbar.dart';
