+++ lib/screens/settings/app_settings/auto_refresh_time_screen.dart
